Moved repomd parser implementation out of the way
